+++ workspaces/x2a/packages/app/e2e-tests/app.test.ts
@@ -20,15 +20,3 @@ import { test, expect } from '@playwright/test';
 test('noop test', async () => {
   expect(true).toBe(true);
 });
-
-/*
-test('App should render the welcome page', async ({ page }) => {
-  await page.goto('/');
-
-  const enterButton = page.getByRole('button', { name: 'Enter' });
-  await expect(enterButton).toBeVisible();
-  await enterButton.click();
-
-  await expect(page.getByText('My Company Catalog')).toBeVisible();
-});
-*/