+++ workspaces/x2a/plugins/x2a/src/components/tools/index.ts
@@ -23,6 +23,7 @@ export * from './getLastPhaseReached';
 export * from './getNextPhase';
 export * from './canRunNextPhase';
 export * from './canCancelPhase';
+export * from './downloadLogFile';
 export * from './hasPhasePrerequisites';
 export * from './areEligibleModulesToRun';
 export * from './isEligibleForRetriggerInit';