+++ workspaces/konflux/plugins/konflux-common/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @red-hat-developer-hub/backstage-plugin-konflux-common
 
+## 0.1.6
+
 ## 0.1.5
 
 ## 0.1.4